Stock_Market_Automation/requirements.txt
fastapi
uvicorn
sqlalchemy
apscheduler
python-dotenv
pydantic
pydantic-settings
requests
pyotp
pytz
pandas
smartapi-python
logzero
websocket